+++ lib/parse.dart
@@ -4,22 +4,19 @@ import 'dart:async';
 import 'dart:convert';
 import 'dart:io';
 
+import 'package:shared_preferences/shared_preferences.dart';
 import 'package:http/http.dart';
-import 'package:intl/intl.dart';
 import 'package:meta/meta.dart';
 import 'package:web_socket_channel/io.dart';
 
 part 'src/base/parse_constants.dart';
-part 'src/data/parse_data_server.dart';
-part 'src/data/parse_data_user.dart';
-part 'src/enums/parse_enum_function_call.dart';
-part 'src/enums/parse_enum_object_call.dart';
-part 'src/enums/parse_enum_user_call.dart';
+part 'src/data/parse_core_data.dart';
+part 'src/enums/parse_enum_api_rq.dart';
 part 'src/network/parse_http_client.dart';
 part 'src/network/parse_livequery.dart';
 part 'src/network/parse_query.dart';
 part 'src/objects/parse_base.dart';
-part 'src/objects/parse_exception.dart';
+part 'src/objects/parse_error.dart';
 part 'src/objects/parse_function.dart';
 part 'src/objects/parse_geo_point.dart';
 part 'src/objects/parse_object.dart';
@@ -28,9 +25,12 @@ part 'src/objects/parse_user.dart';
 part 'src/utils/parse_utils_date.dart';
 part 'src/utils/parse_utils_objects.dart';
 part 'src/utils/parse_utils.dart';
+part 'src/utils/parse_encoder.dart';
+part 'src/utils/parse_decoder.dart';
+part 'src/utils/parse_logger.dart';
 
 class Parse {
-  ParseDataServer data;
+  ParseCoreData data;
   final ParseHTTPClient client = new ParseHTTPClient();
 
   /// To initialise Parse Server in your application
@@ -47,21 +47,22 @@ class Parse {
   // ```
   Parse initialize(appId, serverUrl,
       {debug, appName, liveQueryUrl, masterKey, sessionId}) {
-    ParseDataServer.init(appId, serverUrl,
+    ParseCoreData.init(appId, serverUrl,
         debug: debug,
         appName: appName,
         liveQueryUrl: liveQueryUrl,
         masterKey: masterKey,
         sessionId: sessionId);
 
-    return _newInstance(ParseDataServer());
+    return _newInstance(ParseCoreData());
   }
 
-  /// Creates a singleton instance of [ParseDataServer] that contains all the server information
-  Parse _newInstance(ParseDataServer data) {
+  /// Creates a singleton instance of [ParseCoreData] that contains all the server information
+  Parse _newInstance(ParseCoreData data) {
     var parse = Parse();
     parse.data = data;
     parse.client.data = data;
+    data.initStorage();
     return parse;
   }
 }