+++ lib/src/enums/parse_enum_function_call.dart
@@ -1,4 +1,0 @@
-part of flutter_parse_sdk;
-
-/// Used to define the API calls made in ParseFunction logs
-enum ParseApiFunction { execute }