+++ lib/src/enums/parse_enum_object_call.dart
@@ -1,4 +1,0 @@
-part of flutter_parse_sdk;
-
-/// Used to define the API calls made in ParseObject logs
-enum ParseApiObject { get, getAll, create, save, query, delete }