+++ lib/src/utils/parse_utils.dart
@@ -1,15 +1,11 @@
 part of flutter_parse_sdk;
 
-/// Checks wether debug is enabled
+/// Checks whether debug is enabled
 ///
 /// Debug can be set in 2 places, one global param in the Parse.initialise, and
-/// then can be overriden class by class
-bool isDebugEnabled(bool debug, ParseHTTPClient _client) {
-  if (debug == null) {
-    _client.data.debug != null ? debug = _client.data.debug : debug = false;
-  } else {
-    return debug;
-  }
-
+/// then can be overidden class by class
+bool isDebugEnabled(ParseHTTPClient _client, {objectLevelDebug: false}) {
+  bool debug = objectLevelDebug;
+  if (ParseCoreData().debug != null) debug = ParseCoreData().debug;
   return debug;
 }